Add files via upload
This file contains an outline of a basic RNN system. We will fill in the backprop and the training functions. From there we will adapt to the NLP domain, likely using a one-hot encoding for a vector x_i. We let p be the probabilities of a class, given by the softmax function. Loss will be -ln(p_c).